backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.1.0

# Database
supabase==2.10.0
asyncpg==0.30.0
sqlalchemy==2.0.36

# AI & LLM
langchain==0.3.13
langchain-openai==0.2.0
langchain-google-genai==2.0.0
langchain-community==0.3.13
langgraph==0.2.45
openai==1.99.6
httpx==0.27.2
google-generativeai>=0.7.0,<0.8.0

# Vector Database & Embeddings (Upstash Vector via REST API)
# Note: Using cloud-based embeddings (Gemini/OpenAI API) instead of local models
# This avoids 2GB+ of PyTorch/CUDA dependencies

# Graph Processing
networkx==3.4.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.2.1

# HTTP & WebSockets
websockets==14.1
aiohttp==3.11.11

# Utilities
python-dateutil==2.9.0
tiktoken==0.8.0
tenacity==9.0.0

# Caching
redis==5.0.1

# HTML Parsing
beautifulsoup4==4.12.3

# Payment Processing
razorpay==1.4.2
setuptools>=65.0.0
gunicorn==21.2.0
